+++ client/src/api/applications.js
@@ -1,4 +1,4 @@
-import { patch, remove } from './base'
+import { get, patch, remove } from './base'
 
 export async function acceptApplication (groupId, userId) {
   return await patch(`/applications/${groupId}/${userId}`)
@@ -6,4 +6,8 @@ export async function acceptApplication (groupId, userId) {
 
 export async function declineApplication (groupId, userId) {
   return await remove(`/applications/${groupId}/${userId}`)
+}
+
+export async function getApplications (groupId) {
+  return await get(`/applications/${groupId}`)
 }