+++ src/Modules/Region/Region.js
@@ -25,6 +25,7 @@ import { leaveRegion } from '@/api/regions'
 import '../WallPost/WallPost.css'
 import { initWall } from '@/wall'
 import NewThread from './components/NewThread.vue'
+import ApplicationsList from './components/ApplicationsList'
 
 $(document).ready(() => {
   $('a[href=\'#signout\']').on('click', function () {
@@ -111,5 +112,10 @@ $(document).ready(() => {
       EventList,
     })
     vueApply('#vue-eventlist')
+  } else if (GET('sub') === 'applications') {
+    vueRegister({
+      ApplicationsList,
+    })
+    vueApply('#applications-list')
   }
 })